+++ python/service/dictation_settings.py
@@ -21,17 +21,9 @@ class DictationSettings:
         return self.args.interim_results
 
     def timeouts_map(self):
-
-        # timeouts
-        rec_timeout = (1.0 * self.args.recognition_timeout) / 1000.0
-        comp_timeout = (1.0 * self.args.speech_complete_timeout) / 1000.0
-
-        timeouts = {
+        return {
             "no-input-timeout": str(self.args.no_input_timeout),
-            "endpoint.rule5.min-utterance-length": str(rec_timeout),
-            "endpoint.rule2.min-trailing-silence": str(comp_timeout),
-            "endpoint.rule3.min-trailing-silence": str(comp_timeout),
-            "endpoint.rule4.min-trailing-silence": str(comp_timeout),
+            "speech-complete-timeout": str(self.args.speech_complete_timeout),
+            "speech-incomplete-timeout": str(self.args.speech_incomplete_timeout),
+            "recognition-timeout": str(self.args.recognition_timeout),
         }
-
-        return timeouts