+++ cpp/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Dictation ASR gRPC C++ client Changelog
 
+## [2.0.0] - 2018-09-28
+### Added
+- Returning recognition result finalization cause to API.
+
+### Removed
+- `NO_INPUT_TIMEOUT` speech event type from API.
+
 ## [1.0.2] - 2018-03-23
 ### Changed
 - `DictationSessionConfig` to have fields for sample rate, encoding, language code and to store service settings in a map instead of a string.