+++ python/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Dictation ASR gRPC Python client Changelog
 
+## [2.0.0] - 2018-09-28
+### Added
+- Returning recognition result finalization cause to API.
+- Support for speech incomplete timeout setting.
+
+### Removed
+- `NO_INPUT_TIMEOUT` speech event type from API.
+
 ## [1.1.0] - 2018-07-06
 ### Added
 - Microphone as an audio source support.